pyproject.toml
[project]
name = "employee-ai"
version = "0.3.0"
description = "Personal AI Employee — Gold Tier (Hackathon 0)"
requires-python = ">=3.13"
dependencies = [
    # Bronze Tier
    "watchdog>=4.0.2",
    "python-dotenv>=1.0.0",
    # Silver Tier
    "schedule>=1.2.0",
    "pyyaml>=6.0.1",
    # Gmail Watcher (optional — install separately if needed)
    # "google-auth>=2.27.0",
    # "google-auth-oauthlib>=1.2.0",
    # "google-api-python-client>=2.118.0",
]

[project.optional-dependencies]
gmail = [
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.118.0",
]
# Gold Tier: WhatsApp + social automation via Playwright
playwright = [
    "playwright>=1.43.0",
]
# Gold Tier: full install (all optional deps)
gold = [
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.118.0",
    "playwright>=1.43.0",
]

[project.scripts]
filesystem-watcher  = "watchers.filesystem_watcher:main"
gmail-watcher       = "watchers.gmail_watcher:main"
approval-watcher    = "watchers.approval_watcher:main"
whatsapp-watcher    = "watchers.whatsapp_watcher:main"
linkedin-poster     = "linkedin.post_generator:main"
orchestrator        = "orchestrator:main"
scheduler           = "scheduler:main"
ralph-wiggum        = "ralph_wiggum:main"
weekly-audit        = "audit.weekly_audit:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["watchers", "linkedin", "mcp_servers", "odoo", "social", "audit", "recovery", "hooks"]
